requirements.txt
affine==2.3.1
appdirs==1.4.4
APScheduler==3.9.1
asciitree==0.3.3
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==2.2.0
configobj==5.0.6
dask==2022.10.0
donfig==0.7.0
entrypoints==0.4
eumdac==2.0.0
fasteners==0.18
fsspec==2022.8.2
idna==3.4
importlib-metadata==5.0.0
iso8601==1.1.0
locket==1.0.0
numcodecs==0.10.2
numpy==1.23.4
packaging==21.3
pandas==1.5.0
partd==1.3.0
Pillow==9.2.0
pooch==1.6.0
pykdtree==1.3.5
pyorbital==1.7.3
pyparsing==3.0.9
pyproj==3.4.0
pyresample==1.25.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
rasterio==1.3.2
requests==2.28.1
rioxarray==0.12.2
satpy==0.37.1
scipy==1.9.2
Shapely==1.8.5.post1
six==1.16.0
snuggs==1.4.7
toolz==0.12.0
trollimage==1.18.3
trollsift==0.4.0
typing_extensions==4.4.0
tzdata==2022.5
tzlocal==4.2
urllib3==1.26.12
xarray==2022.10.0
zarr==2.13.3
zipp==3.9.0
